feat: refactor section 2 data fetch into hook
